+++ crates/codegen/src/lib.rs
@@ -6,7 +6,6 @@ mod funcs;
 mod generate_command;
 mod template;
 mod template_file;
-mod templates_command;
 
 pub use asset_file::*;
 pub use codegen::*;
@@ -14,5 +13,4 @@ pub use codegen_error::*;
 pub use generate_command::*;
 pub use template::*;
 pub use template_file::*;
-pub use templates_command::*;
 pub use tera::Context as TemplateContext;